refactor: initial typescript setup
