+++ .eslintrc.json
@@ -6,13 +6,21 @@
     "plugin:react/jsx-runtime",
     "plugin:react-hooks/recommended",
     "plugin:storybook/recommended",
+    "plugin:@typescript-eslint/recommended",
     "prettier"
   ],
   "globals": { "google": "readonly" },
   "rules": {
+    "react/jsx-filename-extension": [
+      2,
+      {
+        "extensions": [".jsx", ".tsx"]
+      }
+    ],
     "react/prop-types": 0,
     "no-console": 1,
     "no-unused-vars": 0,
+    "@typescript-eslint/no-unused-vars": 0,
     "no-param-reassign": 0,
     "react/function-component-definition": [
       2,
