+++ package.json
@@ -37,6 +37,7 @@
   ],
   "devDependencies": {
     "@chromatic-com/storybook": "^1.8.0",
+    "@eslint/js": "^9.21.0",
     "@storybook/addon-essentials": "^8.2.9",
     "@storybook/addon-interactions": "^8.2.9",
     "@storybook/addon-links": "^8.2.9",
@@ -45,6 +46,9 @@
     "@storybook/react": "^8.2.9",
     "@storybook/react-vite": "^8.4.2",
     "@storybook/test": "^8.2.9",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "@typescript-eslint/eslint-plugin": "^8.25.0",
     "@vitejs/plugin-react": "^4.3.1",
     "eslint": "^8.13.0",
     "eslint-config-airbnb": "^19.0.4",
@@ -57,6 +61,8 @@
     "pre-commit": "^1.2.2",
     "prettier": "^2.6.2",
     "storybook": "^8.2.9",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.25.0",
     "vite": "^5.4.12",
     "vite-plugin-eslint": "^1.8.1",
     "vite-plugin-svgr": "^4.2.0",
@@ -71,7 +77,7 @@
     "test:mobile": "cypress run --config-file cypress.mobile.config.js",
     "generate-icons": "npx @svgr/cli --config-file=.svgrrc.json ./src/assets/icons --out-dir=./src/icons && yarn lint --fix",
     "cypress": "cypress open",
-    "lint": "eslint \"**/*.{js,jsx}\" --quiet",
+    "lint": "eslint \"**/*.{js,jsx,ts,tsx}\" --quiet",
     "format": "prettier \"src/**/*.{js,css,html}\" --write",
     "serve": "vite preview",
     "storybook": "storybook dev -p 6006",
