+++ src/index.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import 'bootstrap/dist/js/bootstrap.bundle.min';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import App from 'App';
+import App from 'App.tsx';
 
 let basepath = '';
 // Test-specific routing logic
@@ -18,7 +18,8 @@ if (path) {
 }
 
 const rootElement = document.getElementById('root');
-const root = createRoot(rootElement);
+// We use a non-null assertion here because we are sure that an element with id root exists
+const root = createRoot(rootElement!);
 
 root.render(
   <StrictMode>
